tests/wpt/web-platform-tests/css/css-scroll-snap/snap-to-visible-areas-margin-x-axis.html

<!DOCTYPE html>
<title>
  Snap to an area where the element's scroll-margin is visible but not the
  element itself (using x-axis snap type)
</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#snap-scope"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
div {
  position: absolute;
  margin: 0px;
}

#scroller {
  height: 600px;
  width: 600px;
  overflow: scroll;
  scroll-snap-type: x mandatory;
}

#space {
  width: 2000px;
  height: 2000px;
}

.snap {
  width: 200px;
  height: 200px;
  background-color: blue;
  scroll-snap-align: start;
}

#left-visible {
  left: 0px;
  top: 0px;
}

#right-visible {
  left: 800px;
  top: 0px;
}

#middle-margin-visible {
  left: 700px;
  top: 800px;
  /* 300px makes snap area visible with margin but non-visible without it */
  scroll-margin-top: 300px;
}

</style>
<div id="scroller">
  <div id="space"></div>
  <div id="left-visible" class="snap"></div>
  <div id="middle-margin-visible" class="snap"></div>
  <div id="right-visible" class="snap"></div>
</div>
<script>
test(() => {
  const scroller = document.getElementById("scroller");
  scroller.scrollTo(0, 0);
  assert_equals(scroller.scrollLeft, 0);
  assert_equals(scroller.scrollTop, 0);
  scroller.scrollTo(500, 0);
  assert_equals(scroller.scrollLeft, 700);
  assert_equals(scroller.scrollTop, 0);
}, 'Scroll margin should be considered when calculating snap area visibilty \
while snapping on the x-axis');
</script>
